requirements.txt
python-dotenv
pandas
spotipy